Added State Transitions for Questions and Error Control

+++ src/main/java/metricapp/service/controller/QuestionCRUDController.java
@@ -8,12 +8,16 @@ import org.springframework.stereotype.Service;
 import metricapp.dto.question.QuestionCrudDTO;
 import metricapp.dto.question.QuestionDTO;
 import metricapp.entity.Entity;
+import metricapp.entity.State;
 import metricapp.entity.question.Question;
 import metricapp.exception.BadInputException;
+import metricapp.exception.DBException;
+import metricapp.exception.IllegalStateTransitionException;
 import metricapp.exception.NotFoundException;
 import metricapp.service.spec.ModelMapperFactoryInterface;
 import metricapp.service.spec.controller.QuestionCRUDInterface;
 import metricapp.service.spec.repository.QuestionRepository;
+import metricapp.utility.stateTransitionUtils.AbstractStateTransitionFactory;
 
 @Service
 public class QuestionCRUDController implements QuestionCRUDInterface {
@@ -25,30 +29,37 @@ public class QuestionCRUDController implements QuestionCRUDInterface {
 	private ModelMapperFactoryInterface modelMapperFactory; 
 	
 	@Override
-	public QuestionCrudDTO getQuestionById(String id) {
+	public QuestionCrudDTO getQuestionById(String id) throws NotFoundException, BadInputException {
+		
+		if(id == null){
+			throw new BadInputException("Id cannot be null");
+		}
+		
 		Question question = questionRepository.findQuestionById(id);
+		QuestionCrudDTO questionCrudDTO = new QuestionCrudDTO();
 		try{
-			QuestionCrudDTO questionCrudDTO = new QuestionCrudDTO();
-			
 			questionCrudDTO.addQuestionToList(modelMapperFactory.getLooseModelMapper().map(question, QuestionDTO.class));
 			return questionCrudDTO;
 		}
 		catch(IllegalArgumentException e){
-			System.err.println("No Questions found");
+			throw new NotFoundException("No Questions found");
 		}
-		
-		return null;
 	}
 
 	@Override
-	public QuestionCrudDTO getQuestionByCreatorId(String id) {
+	public QuestionCrudDTO getQuestionByCreatorId(String id) throws BadInputException, NotFoundException {
+		
+		if(id == null){
+			throw new BadInputException("id cannot be null");
+		}
+		
 		ArrayList<Question> questionList = questionRepository.findQuestionByCreatorId(id);
 		
 		QuestionCrudDTO questionCrudDTO = new QuestionCrudDTO();
 		Iterator<Question> questionIter = questionList.iterator();
 		
 		if(!questionIter.hasNext()){
-			return null;
+			throw new NotFoundException("No Questions found for questioner " + id);
 		}
 		
 		while(questionIter.hasNext()){
@@ -59,14 +70,14 @@ public class QuestionCRUDController implements QuestionCRUDInterface {
 	}
 	
 	@Override
-	public QuestionCrudDTO getQuestionByFocus(String focus) {
+	public QuestionCrudDTO getQuestionByFocus(String focus) throws NotFoundException {
 		ArrayList<Question> questionList = questionRepository.findQuestionByFocus(focus);
 		
 		QuestionCrudDTO questionCrudDTO = new QuestionCrudDTO();
 		Iterator<Question> questionIter = questionList.iterator();
 		
 		if(!questionIter.hasNext()){
-			return null;
+			throw new NotFoundException("No question for this focus");
 		}
 		
 		while(questionIter.hasNext()){
@@ -77,14 +88,14 @@ public class QuestionCRUDController implements QuestionCRUDInterface {
 	}
 	
 	@Override
-	public QuestionCrudDTO getQuestionBySubject(String subject) {
+	public QuestionCrudDTO getQuestionBySubject(String subject) throws NotFoundException {
 		ArrayList<Question> questionList = questionRepository.findQuestionBySubject(subject);
 		
 		QuestionCrudDTO questionCrudDTO = new QuestionCrudDTO();
 		Iterator<Question> questionIter = questionList.iterator();
 		
 		if(!questionIter.hasNext()){
-			return null;
+			throw new NotFoundException("No question for this subject");
 		}
 		
 		while(questionIter.hasNext()){
@@ -122,19 +133,32 @@ public class QuestionCRUDController implements QuestionCRUDInterface {
 	}
 
 	@Override
-	public QuestionCrudDTO updateQuestion(QuestionDTO questionDTO) throws BadInputException, NotFoundException {
+	public QuestionCrudDTO updateQuestion(QuestionDTO questionDTO) 
+			throws BadInputException, NotFoundException, IllegalStateTransitionException, DBException{
 		
 		if(questionDTO.getMetadata().getId() == null){
 			throw new BadInputException("Id cannot be null");
 		}
 		
 		questionDTO.getMetadata().setLastVersionDate(LocalDate.now());
-	
+		
 		Question updatedQuestion = modelMapperFactory.getLooseModelMapper().map(questionDTO, Question.class);
+		
+		Question oldQuestion = questionRepository.findQuestionById(questionDTO.getMetadata().getId());
+		
+		System.out.println(oldQuestion.getState());
+		System.out.println(updatedQuestion.getState());
+		
+		stateTransition(oldQuestion, updatedQuestion);
 		
 		if(questionRepository.exists(updatedQuestion.getId())){
 			QuestionCrudDTO questionCrudDTO = new QuestionCrudDTO();
-			questionRepository.save(updatedQuestion);
+			try{
+				questionRepository.save(updatedQuestion);
+			}
+			catch(Exception e){
+				throw new DBException("Error in saving question in repository");
+			}
 			
 			questionCrudDTO.addQuestionToList(questionDTO);
 			return questionCrudDTO;
@@ -161,13 +185,18 @@ public class QuestionCRUDController implements QuestionCRUDInterface {
 	}
 
 	@Override
-	public boolean deleteQuestionById(String id) {
+	public void deleteQuestionById(String id) throws IllegalStateTransitionException, NotFoundException {
+				
 		if(questionRepository.exists(id)){
+			if (!questionRepository.findQuestionById(id).getState().equals(State.Suspended)) {
+				System.out.println(questionRepository.findQuestionById(id).getState());
+				throw new IllegalStateTransitionException("A question must be Suspended before deletion");
+			}
+
 			questionRepository.delete(id);
-			return true;
 		}
 		else{
-			return false;
+			throw new NotFoundException("The question does not exists");
 		}
 
 	}
@@ -175,8 +204,15 @@ public class QuestionCRUDController implements QuestionCRUDInterface {
 	@Override
 	public void deleteAllQuestions() {
 		questionRepository.deleteAll();
+	}
+	
+	private void stateTransition(Question oldQuestion, Question newQuestion)
+			throws IllegalStateTransitionException, NotFoundException {
+		
+		if (!oldQuestion.getState().equals(newQuestion.getState())) {
+			AbstractStateTransitionFactory.getFactory(Entity.Question).transition(oldQuestion, newQuestion).execute();
+	 	}
 		
 	}
-
 
 }
